Update coding style

+++ md4.h
@@ -23,22 +23,21 @@
 #include <stddef.h>
 #include <stdint.h>
 
-typedef struct
-{
-	size_t    len;
-	size_t   bufLen;
-	uint8_t  buffer[64];
-	uint32_t A;
-	uint32_t B;
-	uint32_t C;
-	uint32_t D;
+typedef struct {
+    size_t len;
+    size_t bufLen;
+    uint8_t buffer[64];
+    uint32_t A;
+    uint32_t B;
+    uint32_t C;
+    uint32_t D;
 } MD4_CTX;
 
-void MD4Init  (MD4_CTX* md4);
-void MD4Block (MD4_CTX* md4, const uint8_t block[64]);
-void MD4Update(MD4_CTX* md4, const uint8_t* data, uint64_t len);
-void MD4Final (MD4_CTX* md4, uint8_t dst[16]);
+void MD4Init(MD4_CTX* md4);
+void MD4Block(MD4_CTX* md4, const uint8_t block[64]);
+void MD4Update(MD4_CTX* md4, const uint8_t* data, size_t len);
+void MD4Final(MD4_CTX* md4, uint8_t dst[16]);
 
-void MD4(uint8_t dst[16], const uint8_t* src, uint64_t slen);
+void MD4(uint8_t dst[16], const uint8_t* src, size_t slen);
 
 #endif
