+++ rijndael.h
@@ -23,7 +23,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-void Rijndael(const uint8_t* key, const uint8_t* in, uint8_t* out, bool inverse, size_t Nk, size_t Nr);
+void Rijndael(const uint8_t* key, const uint8_t* in, uint8_t* out, bool inverse, int Nk, int Nr);
 
 void Rijndael128(const uint8_t key[16], const uint8_t in[16], uint8_t out[16], bool inverse);
 void Rijndael192(const uint8_t key[24], const uint8_t in[16], uint8_t out[16], bool inverse);
